refactor: create PaymentSelectionSectionItem
