+++ src/translations/screens/subscreens/PaymentMethods.ts
@@ -41,11 +41,17 @@ const PaymentMethodsTexts = {
         `Activate to remove ${paymentName} ending in ${masked_pan}`,
         `Aktiver for å fjerne ${paymentName} som sluttar på ${masked_pan}`,
       ),
-    editCardIcon: (paymentName: string, masked_pan: string) =>
+    editCard: (paymentName: string) =>
       _(
-        `Aktiver for å redigere betalingsmetode ${paymentName} som slutter på ${masked_pan}`,
-        `Activate to edit payment method ${paymentName} ending in ${masked_pan}`,
-        `Aktiver for å endre betalingsmåte ${paymentName} som sluttar på ${masked_pan}`,
+        `Betalingsmetode ${paymentName}. Aktiver for å endre betalingsmetode.`,
+        `Payment method ${paymentName}. Activate to edit payment method.`,
+        `Betalingsmåte ${paymentName}. Aktiver for å endre betalingsmåte.`,
+      ),
+    editCardWithMaskedPan: (paymentName: string, masked_pan: string) =>
+      _(
+        `Betalingsmetode ${paymentName} som slutter på ${masked_pan}. Aktiver for å endre betalingsmetode.`,
+        `Payment method ${paymentName} ending in ${masked_pan}. Activate to edit payment method.`,
+        `Betalingsmåte ${paymentName} som sluttar på ${masked_pan}. Aktiver for å endre betalingsmåte.`,
       ),
   },
   addPaymentMethod: _(
